backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiosqlite
python-dotenv
ebooklib
beautifulsoup4
pymupdf
google-genai
python-multipart
aiofiles
pytest
pytest-asyncio
httpx